+++ lib/utils/app_router.dart
@@ -1,14 +1,14 @@
 import 'package:go_router/go_router.dart';
 import 'package:nutripic/components/bottom_navigator_bar.dart';
 import 'package:nutripic/models/user_model.dart';
-import 'package:nutripic/view_models/login_view_model.dart';
-import 'package:nutripic/view_models/signup_view_model.dart';
+import 'package:nutripic/view_models/login/login_view_model.dart';
+import 'package:nutripic/view_models/login/signup_view_model.dart';
 import 'package:nutripic/view_models/user_info/user_edit_view_model.dart';
 import 'package:nutripic/view_models/user_info/user_info_view_model.dart';
-import 'package:nutripic/views/signup_view.dart';
+import 'package:nutripic/views/login/signup_view.dart';
 import 'package:nutripic/views/camera_view.dart';
 import 'package:nutripic/views/diary_view.dart';
-import 'package:nutripic/views/login_view.dart';
+import 'package:nutripic/views/login/login_view.dart';
 import 'package:nutripic/views/recipe_view.dart';
 import 'package:nutripic/views/refrigerator_view.dart';
 import 'package:nutripic/views/user_info/user_edit_view.dart';
