+++ lib/views/login/login_view.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:nutripic/view_models/login_view_model.dart';
+import 'package:nutripic/view_models/login/login_view_model.dart';
 
 class LoginView extends StatefulWidget {
   const LoginView({super.key});
